feat: VisaDocument + user verification fields + migration
