+++ extensions.py
@@ -1,5 +1,7 @@
+from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
 
 # Shared SQLAlchemy instance for the application
 # Imported by modules that require database access.
 db = SQLAlchemy()
+migrate = Migrate()
