+++ models/base.py
@@ -1,0 +1,12 @@
+from datetime import datetime
+
+from extensions import db
+
+
+class TimestampMixin:
+    """Mixin providing created/updated timestamps."""
+
+    created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
+    updated_at = db.Column(
+        db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow, nullable=False
+    )
