+++ requirements.txt
@@ -1,4 +1,5 @@
 Flask==2.3.2
+Flask-Migrate==4.0.4
 Flask-SQLAlchemy==3.0.5
 PyJWT==2.8.0
 pytest==7.4.2
